server/src/main/java/de/zalando/zally/rule/zalando/AvoidTrailingSlashesRule.kt
package de.zalando.zally.rule.zalando

import de.zalando.zally.rule.ApiAdapter
import de.zalando.zally.rule.api.Check
import de.zalando.zally.rule.api.Rule
import de.zalando.zally.rule.api.Severity
import de.zalando.zally.rule.api.Violation
import de.zalando.zally.util.PatternUtil

@Rule(
        ruleSet = ZalandoRuleSet::class,
        id = "136",
        severity = Severity.MUST,
        title = "Avoid Trailing Slashes"
)
class AvoidTrailingSlashesRule {
    private val description = "Rule avoid trailing slashes is not followed"

    @Check(severity = Severity.MUST)
    fun validate(adapter: ApiAdapter): Violation? {
        val paths = adapter.openAPI.paths.orEmpty().keys.filter { it != null && PatternUtil.hasTrailingSlash(it) }
        return if (!paths.isEmpty()) Violation(description, paths) else null
    }
}
